third_party/blink/web_tests/platform/win11-arm64/http/tests/eventsource/workers/eventsource-cors-redirect-expected.txt

CONSOLE ERROR: EventSource's response has a MIME type ("text/html") that is not "text/event-stream". Aborting the connection.
[Worker] Test that basic EventSource cross-origin requests fail on redirect.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../script-tests/eventsource-cors-redirect.js
PASS [Worker] es = new EventSource("/resources/redirect.php?code=307&url=http://127.0.0.1:8080/eventsource/resources/es-cors-basic.php") did not throw exception.
PASS [Worker] es.readyState is EventSource.CLOSED
PASS successfullyParsed is true

TEST COMPLETE
